fix(trusty-search): split the drop-reporting tests out of tests_search.rs
tests_search.rs matches none of the test-file name patterns, so it is capped
at 500 production SLOC; the two new tests pushed it to 528. Splitting inside
the PR that trips the cap, per the no-standalone-cap-fix rule.
